Make `EscTag` a `ParserTag`
